+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Improvements
 ### MLEM.Ui
 Additions
 - Added some extension methods for querying Anchor types
+- Added Element.AutoSizeAddedAbsolute to allow for more granular control of auto-sizing
 
 Improvements
 - Allow elements to auto-adjust their size even when their children are aligned oddly